% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_Iterator.R
\name{as_Iterator}
\alias{as_Iterator}
\title{Coerce a list to Iterator object}
\usage{
as_Iterator(list)
}
\arguments{
\item{list}{a list to coerce to Iterator}
}
\description{
Coerce a list to Iterator object
}
