% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_along.R
\name{char_along}
\alias{char_along}
\title{Simple wrapper to aid in iterating \code{iterstring}s}
\usage{
char_along(string)
}
\arguments{
\item{string}{a string to create an integer sequence from}
}
\description{
\code{char_along(string)} simply means \code{seq_len(nchar(string))}
}
