% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{current}
\alias{current}
\title{Get the current value of an Iterator without changing its state}
\usage{
current(iter)
}
\arguments{
\item{iter}{An \code{Iterator} object}
}
\value{
The current value of \code{iter}
}
\description{
An \code{Iterator} \code{yield}s a variable every time \code{yield_next()} is called.
Get the current value of that variable without changing the state of the Iterator.
}
