% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islist.R
\name{islist}
\alias{islist}
\alias{as.character.islist}
\alias{as_islist}
\title{Functions to create and manipulate \code{islist}s, lists of \code{iterstring}s}
\usage{
\method{as.character}{islist}(islist)

as_islist(object)
}
\arguments{
\item{islist}{an \code{islist} to turn into a character vector}

\item{object}{an object to convert to \code{islist}}
}
\description{
Since an \code{iterstring} is an S3 object, there can be no atomic vectors of \code{iterstrings}.
Therefore containers of \code{iterstring}s are \code{list}s, and their S3 class is \code{islist}.
An \code{islist} is just a list of \code{iterstring}s, so it is inadvisable to try to call
\code{as_islist} on any object except a character vector, a list of \code{iterstring}s,
or a list of character vectors each of length one.
}
