% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterstring.R
\name{iterstring}
\alias{iterstring}
\title{Making an iterable string}
\usage{
iterstring(string)
}
\arguments{
\item{string}{a string to make iterable}
}
\value{
an iterable object of S3 class \code{iterstring}
}
\description{
In R, string objects are just elements of character vectors,
so they are not themselves iterable. An \code{iterstring} object changes that.
Mostly, \code{iterstrings} behave like Python strings, with some \code{R}-like exceptions
}
