% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.Iterator}
\alias{print.Iterator}
\title{Summary and print method for objects of S3 class Iterator}
\usage{
\method{print}{Iterator}(iter)
}
\arguments{
\item{iter}{an object of S3 class Iterator}
}
\description{
Summary and print method for objects of S3 class Iterator
}
\examples{
expr <- 'if (n \%\% 2 == 0) n <- n / 2 else n <- n*3 + 1'
collatz <- Iterator(result = expr,
                     initial = c(n = 50),
                     yield = n)
print(collatz)
summary(collatz) #equivalent
}
