% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Use an approximation of Python's set builder}
\usage{
range(a, b)
}
\arguments{
\item{a}{minimum, integer}

\item{b}{supremum, integer}
}
\description{
Notation in R, based around the magrittr pipe
}
\examples{
range(1,5)
range(9,10)
}
