% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{Extract}
\alias{Extract}
\alias{[.iterstring}
\alias{[[.iterstring}
\alias{[<-.iterstring}
\alias{[[<-.iterstring}
\title{Get and set elements of an \code{iterstring}}
\usage{
\method{[}{iterstring}(string, range)

\method{[[}{iterstring}(string, index)

\method{[}{iterstring}(string, range) <- value

\method{[[}{iterstring}(string, index) <- value
}
\arguments{
\item{string}{an \code{iterstring} to subset}

\item{range}{a range of indices to extract with \code{[}}

\item{index}{an index to extract with \code{[[}}

\item{value}{a suibable replacement value, must be a single character with \code{[[}}
}
\description{
Get and set elements of a string, similar to an R vector or a Python string.
Note that indices start at 1 like R vectors, not Python strings.
Includes \code{slice}-ing vectors with positive or negative indices, like in Python.
Follows R convention of \code{[} single-bracket subsetting multiple elements and preserving
attributes, and \code{[[} double bracket subsetting a single element (in this case a single character)
and stripping attributes.

Note how negative indices work—the last character of a string \code{str} is \code{str[-1]},
the second to last is \code{str[-2]}, etc. \code{str[[0]]} will throw an error, as will \code{str[[-0]]}.
\code{str[0]} and \code{str[-0]} will return an empty string.
}
