% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_next.R
\name{yield_next}
\alias{yield_next}
\title{Equivalent to Python's 'next'}
\usage{
yield_next(gen)
}
\arguments{
\item{gen}{A generator object}
}
\value{
An object of whatever type 'result' evaluates to from the Generator
}
\description{
Finds the value of the next iteration of 'gen'.
If argument is list, attempts to coerce to Generator.
}
