% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/that_for_{all,any}.R
\name{sets}
\alias{sets}
\alias{that_for_all}
\alias{that_for_any}
\alias{we_have}
\title{R Set Comprehension}
\usage{
that_for_all(set1, set2)

that_for_any(set1, set2)

we_have(that_for, formula)
}
\arguments{
\item{set1}{A superset}

\item{set2}{A subset of set1}

\item{formula}{A logical formula to test elements from set1 against those from set2}
}
\value{
For that_for_all and that_for_any, an object of S3 class that_for_all or that_for_any.
For we_have, a vector of the same type as set1
}
\description{
Replicates Python's set comprehension using tidy syntax.
Always use the syntax: set1 \%>\% that_for_all(set2) \%>\% we_have(formula).
}
\examples{
1:10 \%>\% that_for_all(1:5) \%>\% we_have(~.x \%\% .y != 0)
c("I", "Don't", "wan't", "chicken") \%>\% that_for_all('\'') \%>\% we_have(~stringr::str_detect(.x, .y))

}
