% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generator.R
\name{Generator}
\alias{Generator}
\title{Making a Python-style Generator
Uses R's S3 OO paradigm and lazy evaluation}
\usage{
Generator(result, current, yield)
}
\arguments{
\item{result}{R expression to run}

\item{current}{current value}

\item{yield}{variable to yield}
}
\description{
Making a Python-style Generator
Uses R's S3 OO paradigm and lazy evaluation
}
\examples{
expr <- "if (i == 0) {
            n0 <- 0;
           n1 <- 0;
           n <- n0 + n1; };
        if (i == 1) {
           n0 <- 0; n1 <- 1;
           n <- n0 + n1;
        };
        if (i > 1) {

           n <- n0 + n1;
           a <- n1;
            b <- n0 + n1;
            n0 <- a;
            n1 <- b;
          };
          i <- i + 1;"

gen <- Generator(result = expr,
                current = c(i = 0,
                            a = 0,
                            b = 0,
                            n0 = 0,
                            n1 = 0,
                            n = 0),
                yield = n)
}
