% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\alias{move_next}
\alias{move_more}
\title{Increment an Iterator without returning the Value(s)}
\usage{
move_next(iter)

move_more(iter, more = 1L)
}
\arguments{
\item{iter}{An Iterator object object}

\item{more}{How many times to iterate}
}
\description{
Increments the Iterator without returning anything.
}
