% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Create a Python-style range vector}
\usage{
range(a, b)
}
\arguments{
\item{a}{minimum, integer}

\item{b}{supremum, integer}
}
\description{
Create a Python-style range vector
}
\examples{
range(1,5)
range(9,10)
}
