% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Use an approximation of Python's set builder
Notation in R, based around the magrittr pipe}
\usage{
range(a, b)
}
\arguments{
\item{a}{desc}

\item{b}{desc}
}
\description{
Use an approximation of Python's set builder
Notation in R, based around the magrittr pipe
}
