% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_next.R
\name{yield_next}
\alias{yield_next}
\title{Equivalent to Python's 'next'
Finds the value of the next iteration of 'gen'
If argument is list, attempts to coerce to Generator}
\usage{
yield_next(gen)
}
\arguments{
\item{gen}{generator object}
}
\description{
Equivalent to Python's 'next'
Finds the value of the next iteration of 'gen'
If argument is list, attempts to coerce to Generator
}
