% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_next.R
\name{yield_next}
\alias{yield_next}
\title{Increment an Iterator and Return the Next Value}
\usage{
yield_next(iter)
}
\arguments{
\item{iter}{An Iterator object object}
}
\value{
An object of whatever type 'result' evaluates to from the Iterator
}
\description{
Finds the value of the next iteration of an Iterator object
and increments the Iterator to the next value.
}
