% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_next.R
\name{yield}
\alias{yield}
\alias{yield_next}
\alias{yield_more}
\title{Increment an Iterator and Return the Next Value(s)}
\usage{
yield_next(iter)

yield_more(iter, more = 1L)
}
\arguments{
\item{iter}{An Iterator object}

\item{more}{How many values to yield}
}
\value{
An object of whatever type \code{result} evaluates to from the Iterator, or
a vector of that type in the case of \code{yield_more(iter, more > 1L)}.
}
\description{
Finds the value of the next iteration(s) of an Iterator object
and increments the Iterator to the next value(s).
}
